revision: add rdiff_other_arg to rev_info
git-format-patch(1) needs to pass `--[no-]notes` options on to the
range-diff subprocess in `range-diff.c`.  This is handled in `builtin/
log.c` by the local variable `other_arg` in the case of multiple
commits, but not in the single commit case where there is no cover
letter and the range-diff is on that single resulting patch; the
range-diff is then made in `log-tree.c`, whither `other_arg` has not
been propagated.

git-format-patch(1) is supposed to treat Git notes the same between
notes output beneath the commit message and the notes output for the
range-diff.  But this lack of notes handling in `log-tree.c` breaks
that expected behavior; range-diff notes handling for a single patch
acts like a normal git-range-diff(1) invocation with regards to notes.
You can, for example, end up with a patch where the note beneath the
commit message has the correct notes namespace, but the range-diff has
all the notes that are configured to be displayed by git-log(1).[1]

We need to fix this.  But first lay the groundwork by converting
`other_arg` to a struct member; next we can simply use that member
in `log-tree.c` without having to thread it from `builtin/log.c`.

No functional changes.

† 1: See the configuration variable `format.notes` for git-format-
     patch(1); c.f. `notes.displayRef` for git-log(1).  These two
     have nothing to do with each other.

Signed-off-by: Kristoffer Haugsbakk <code@khaugsbakk.name>
Signed-off-by: Junio C Hamano <gitster@pobox.com>